requirements.txt
#sentence-transformers

# Use Conda for CUDA version on a GPU enabled machine
# conda install pytorch torchvision torchaudio pytorch-cuda=12.4 -c pytorch -c nvidia
flash_attn==2.5.8
numpy==1.24.4
Pillow==10.3.0
Requests==2.31.0
torch==2.3.0
torchvision==0.18.0
transformers==4.43.1
accelerate==0.30.0

#-r requirements/internvl_chat.txt
#-r requirements/streamlit_demo.txt
#-r requirements/classification.txt
#-r requirements/segmentation.txt
